@lmb 2025-05-07
@patrickpichler could you give this branch a try and let me know what you think?