@patrickpichler 2025-05-21
Sorry for the super late response. I did run the benchmark again and here are the results:
Without your change
BenchmarkNewCollectionManyProgs-4   	     15	 75442629 ns/op	74477811 B/op	 809763 allocs/op

With your change
BenchmarkNewCollectionManyProgs-4   	    272	  4193884 ns/op	2734829 B/op	  30806 allocs/op

With the optimization I did in the draft PR (#1690)
BenchmarkNewCollectionManyProgs-4 184 5807742 ns/op 4134444 B/op 17325 allocs/op

When performing a single run, the total number of allocated memory according to pprof went down from `~660MB` to `~26MB`, so this is definitely a win. 

I guess there are a few more ways on how to reduce the number of allocations, but for my use case, this should be good enough. 

Thanks a lot!